[refactor]: pagenation 부분 코드 분리
