+++ src/index.html
@@ -21,34 +21,21 @@
   <body>
     <main>
       <h1>My Agora States</h1>
+
       <button class="modalBtn">질문하기</button>
+      <form class="search__box">
+        <input
+          type="text"
+          name="search"
+          id="search"
+          required
+          placeholder="Search title or name"
+        />
+        <button><i class="fa-solid fa-magnifying-glass"></i></button>
+      </form>
 
       <article class="discussion__wrapper">
-        <ul class="discussions__container">
-          <!--           <li class="discussion__container">
-            <div class="discussion__avatar--wrapper">
-              <img
-                class="discussion__avatar--image"
-                src="https://avatars.githubusercontent.com/u/12145019?s=64&u=5c97f25ee02d87898457e23c0e61b884241838e3&v=4"
-                alt="avatar of kimploo"
-              />
-            </div>
-            <div class="discussion__content">
-              <h2 class="discussion__title">
-                <a
-                  target="_blank"
-                  href="https://github.com/codestates-seb/agora-states-fe/discussions/6"
-                  >[notice] 좋은 질문하는 법</a
-                >
-              </h2>
-              <div class="discussion__information">
-                kimploo / 2022-04-22T14:08:33Z
-              </div>
-            </div>
-            <div class="discussion__answered"><p></p></div>
-            
-          </li> -->
-        </ul>
+        <ul class="discussions__container"></ul>
         <div class="buttons"></div>
       </article>
     </main>
@@ -80,3 +67,27 @@
   </body>
   <script src="./js/script.js" type="module"></script>
 </html>
+
+<!--           <li class="discussion__container">
+            <div class="discussion__avatar--wrapper">
+              <img
+                class="discussion__avatar--image"
+                src="https://avatars.githubusercontent.com/u/12145019?s=64&u=5c97f25ee02d87898457e23c0e61b884241838e3&v=4"
+                alt="avatar of kimploo"
+              />
+            </div>
+            <div class="discussion__content">
+              <h2 class="discussion__title">
+                <a
+                  target="_blank"
+                  href="https://github.com/codestates-seb/agora-states-fe/discussions/6"
+                  >[notice] 좋은 질문하는 법</a
+                >
+              </h2>
+              <div class="discussion__information">
+                kimploo / 2022-04-22T14:08:33Z
+              </div>
+            </div>
+            <div class="discussion__answered"><p></p></div>
+            
+          </li> -->
